move unit tests into new submodules
