+++ crates/astria-sequencer/src/action_handler/test_utils.rs
@@ -1,0 +1,5 @@
+use astria_core::primitive::v1::asset;
+
+pub(super) fn test_asset() -> asset::Denom {
+    "test".parse().unwrap()
+}
