--- crates/astria-sequencer/src/action_handler/tests/mod.rs
@@ -1,10 +1,0 @@
-use astria_core::primitive::v1::asset;
-
-mod bridge_sudo_change;
-mod bridge_unlock;
-mod fee_change;
-mod ics20_withdrawal;
-
-fn test_asset() -> asset::Denom {
-    "test".parse().unwrap()
-}
